+++ SUPPORT.md
@@ -2,10 +2,10 @@
 
 ## How to file issues and get help
 
-This project uses GitHub [issues](https://github.com/github/spdx-expression/issues) to track bugs and feature requests. Please search the existing issues before filing new issues to avoid duplicates. If not found, file your bug or feaure request as a new issue.
+This project uses GitHub [issues](https://github.com/github/go-spdx/issues) to track bugs and feature requests. Please search the existing issues before filing new issues to avoid duplicates. If not found, file your bug or feaure request as a new issue.
 
-For help or questions about using this project, please see the project [Discussions](https://github.com/github/spdx-expression/discussions) where you can ask a question in [Q&A](https://github.com/github/spdx-expression/discussions/categories/q-a), propose an idea in [Ideas](https://github.com/github/spdx-expression/discussions/categories/ideas), or share your work in [Show and Tell](https://github.com/github/spdx-expression/discussions/categories/show-and-tell).
+For help or questions about using this project, please see the project [Discussions](https://github.com/github/go-spdx/discussions) where you can ask a question in [Q&A](https://github.com/github/go-spdx/discussions/categories/q-a), propose an idea in [Ideas](https://github.com/github/go-spdx/discussions/categories/ideas), or share your work in [Show and Tell](https://github.com/github/go-spdx/discussions/categories/show-and-tell).
 
-The spdx-expression package is under active development and maintained by GitHub staff.  Community [contributions](./CONTRIBUTING.md) are always welcome. We will do our best to respond to issues, feature requests, and community questions in a timely manner.
+The go-spdx package is under active development and maintained by GitHub staff.  Community [contributions](./CONTRIBUTING.md) are always welcome. We will do our best to respond to issues, feature requests, and community questions in a timely manner.
 
 Support for this project is limited to the resources listed above.
