+++ expression/compare.go
@@ -1,4 +1,4 @@
-package spdxexp
+package expression
 
 func compareGT(first *node, second *node) bool {
 	if !first.isLicense() || !second.isLicense() {
