+++ expression/satisfies.go
@@ -1,4 +1,4 @@
-package spdxexp
+package expression
 
 import (
 	"errors"
