+++ expression/satisfies_test.go
@@ -1,4 +1,4 @@
-package spdxexp
+package expression
 
 import (
 	"errors"
@@ -31,7 +31,7 @@ func TestValidateLicenses(t *testing.T) {
 
 // TestSatisfiesSingle lets you quickly test a single call to Satisfies with a specific license expression and allowed list of licenses.
 // To test a different expression, change the expression, allowed licenses, and expected result in the function body.
-// TO RUN: go test ./spdxexp -run TestSatisfiesSingle
+// TO RUN: go test ./expression -run TestSatisfiesSingle
 func TestSatisfiesSingle(t *testing.T) {
 	// Update these to test a different expression.
 	expression := "BSD-3-Clause AND GPL-2.0"
