+++ expression/test_helper.go
@@ -1,4 +1,4 @@
-package spdxexp
+package expression
 
 // getLicenseNode is a test helper method that is expected to create a valid
 // license node.  Use this function when the test data is known to be a valid
