--- spdxexp/doc.go
@@ -1,7 +1,0 @@
-/*
-Spdxexp package validates licenses and determines if a license expression is satisfied by a list of licenses.
-Validity of a license is determined by the [SPDX license list].
-
-[SPDX license list]: https://spdx.org/licenses/
-*/
-package spdxexp
